# Reduce GC objects count