It reduces the number of objects in the heap.

To test the number of objects in the heap I wrote the following code in the BenchmarkFair func:

```go
	memOld := runtime.MemStats{}
	runtime.ReadMemStats(&memOld)
	defer func() {
		memNew := runtime.MemStats{}
		runtime.ReadMemStats(&memNew)
		b.Logf("HeapObjects=%v; HeapAlloc=%v\n", memNew.HeapObjects-memOld.HeapObjects, memNew.HeapAlloc-memOld.HeapAlloc)
	}()
```

old: `insane_perf_test.go:134: HeapObjects=140189; HeapAlloc=8590888`
new: `insane_perf_test.go:134: HeapObjects=72100; HeapAlloc=7133104`

And it also slightly increases performance:

```
$ benchstat -table .fullname old.txt new.txt
.fullname: Fair/complex-stable-flavor|complex-8
                                     │   old.txt   │              new.txt               │
                                     │   sec/op    │   sec/op     vs base               │
Fair/complex-stable-flavor|complex-8   1.121m ± 1%   1.113m ± 1%  -0.68% (p=0.007 n=10)

                                     │   old.txt    │               new.txt               │
                                     │     B/s      │     B/s       vs base               │
Fair/complex-stable-flavor|complex-8   1.069Gi ± 1%   1.076Gi ± 1%  +0.68% (p=0.007 n=10)

                                     │   old.txt    │            new.txt             │
                                     │     B/op     │     B/op      vs base          │
Fair/complex-stable-flavor|complex-8   1.852Ki ± 1%   1.846Ki ± 1%  ~ (p=0.288 n=10)

                                     │  old.txt   │            new.txt             │
                                     │ allocs/op  │ allocs/op   vs base            │
Fair/complex-stable-flavor|complex-8   32.00 ± 0%   32.00 ± 0%  ~ (p=1.000 n=10) ¹
¹ all samples are equal

.fullname: Fair/complex-chaotic-flavor|complex-8
                                      │   old.txt   │            new.txt            │
                                      │   sec/op    │   sec/op     vs base          │
Fair/complex-chaotic-flavor|complex-8   279.9µ ± 1%   279.4µ ± 1%  ~ (p=0.481 n=10)

                                      │   old.txt    │            new.txt             │
                                      │     B/s      │     B/s       vs base          │
Fair/complex-chaotic-flavor|complex-8   283.1Mi ± 1%   283.7Mi ± 1%  ~ (p=0.481 n=10)

                                      │   old.txt    │             new.txt              │
                                      │     B/op     │     B/op      vs base            │
Fair/complex-chaotic-flavor|complex-8   39.81Ki ± 0%   39.81Ki ± 0%  ~ (p=1.000 n=10) ¹
¹ all samples are equal

                                      │   old.txt   │             new.txt             │
                                      │  allocs/op  │  allocs/op   vs base            │
Fair/complex-chaotic-flavor|complex-8   1.532k ± 0%   1.532k ± 0%  ~ (p=1.000 n=10) ¹
¹ all samples are equal

.fullname: Fair/get-stable-flavor|get-8
                             │   old.txt   │            new.txt            │
                             │   sec/op    │   sec/op     vs base          │
Fair/get-stable-flavor|get-8   440.9n ± 1%   442.1n ± 1%  ~ (p=0.811 n=10)

                             │   old.txt    │            new.txt             │
                             │     B/s      │     B/s       vs base          │
Fair/get-stable-flavor|get-8   1.327Ti ± 1%   1.323Ti ± 1%  ~ (p=0.853 n=10)

                             │  old.txt   │            new.txt             │
                             │    B/op    │    B/op     vs base            │
Fair/get-stable-flavor|get-8   0.000 ± 0%   0.000 ± 0%  ~ (p=1.000 n=10) ¹
¹ all samples are equal

                             │  old.txt   │            new.txt             │
                             │ allocs/op  │ allocs/op   vs base            │
Fair/get-stable-flavor|get-8   0.000 ± 0%   0.000 ± 0%  ~ (p=1.000 n=10) ¹
¹ all samples are equal

.fullname: Fair/get-chaotic-flavor|get-8
                              │   old.txt   │              new.txt               │
                              │   sec/op    │   sec/op     vs base               │
Fair/get-chaotic-flavor|get-8   5.104µ ± 0%   5.015µ ± 0%  -1.75% (p=0.000 n=10)

                              │   old.txt    │               new.txt               │
                              │     B/s      │     B/s       vs base               │
Fair/get-chaotic-flavor|get-8   7.581Gi ± 0%   7.717Gi ± 1%  +1.79% (p=0.000 n=10)

                              │   old.txt    │             new.txt              │
                              │     B/op     │     B/op      vs base            │
Fair/get-chaotic-flavor|get-8   4.443Ki ± 0%   4.443Ki ± 0%  ~ (p=1.000 n=10) ¹
¹ all samples are equal

                              │  old.txt   │            new.txt             │
                              │ allocs/op  │ allocs/op   vs base            │
Fair/get-chaotic-flavor|get-8   168.0 ± 0%   168.0 ± 0%  ~ (p=1.000 n=10) ¹
¹ all samples are equal
```