@vadimalekseev 2024-01-20
Later I will attach the effect on GC in the file.d