@vadimalekseev 2024-01-23
Restart was at 12:00 and 12:15 to make it easier to compare two different versions
New release at 13:14

<img width="1153" alt="image" src="https://github.com/ozontech/insane-json/assets/45711528/f8220a87-811f-4fb7-b618-e419a6ac6dd8">
<img width="1245" alt="image" src="https://github.com/ozontech/insane-json/assets/45711528/783cd590-e659-442d-88b5-a3b745308eac">
<img width="1213" alt="image" src="https://github.com/ozontech/insane-json/assets/45711528/9cce9577-f7cd-4163-89b1-782a188892d9">
<img width="1190" alt="image" src="https://github.com/ozontech/insane-json/assets/45711528/1d9ff9c3-95c0-4264-b46a-247615dca2ea">
<img width="1212" alt="image" src="https://github.com/ozontech/insane-json/assets/45711528/f5403067-f4da-4f4b-b799-dda3292a29b8">
<img width="1202" alt="image" src="https://github.com/ozontech/insane-json/assets/45711528/e90d5946-7991-4f8a-8d6f-dc085d446d5f">
<img width="1239" alt="image" src="https://github.com/ozontech/insane-json/assets/45711528/d7be839f-baa1-4493-a96b-44430ada8ddd">

Screenshots show metrics of one of the most loaded file.d cluster
Other metrics have not changed
E2E and unit tests in file.d passed successfully