@vadimalekseev 2024-01-24
After 24h

<img width="1847" alt="image" src="https://github.com/ozontech/insane-json/assets/45711528/40db52a5-e7ec-43bc-a114-61c9e4349dbf">
<img width="1861" alt="image" src="https://github.com/ozontech/insane-json/assets/45711528/82ee61a4-4f6b-49d4-aa63-071f7ef67473">
<img width="1856" alt="image" src="https://github.com/ozontech/insane-json/assets/45711528/a018ee5b-628f-4ff7-9d63-6a0f29108c9f">
<img width="1855" alt="image" src="https://github.com/ozontech/insane-json/assets/45711528/7bd42d71-f77f-45d4-93d7-137a2dbb88e4">
<img width="1856" alt="image" src="https://github.com/ozontech/insane-json/assets/45711528/a94bd29e-d33b-486d-9237-0f1871c88657">
<img width="1868" alt="image" src="https://github.com/ozontech/insane-json/assets/45711528/3fcfa23c-1922-4459-b186-1c39cb50cf26">